+++ Reminder/ListViewController/ReminderListViewController+DataSource.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+extension ReminderListViewController {
+    typealias DataSource = UICollectionViewDiffableDataSource<Int, String>
+    typealias Snapshot = NSDiffableDataSourceSnapshot<Int, String>
+    
+    func cellRegistrationHandler(cell: UICollectionViewListCell, indexPath: IndexPath, id: String ) {
+        let reminder = Reminder.sampleData[indexPath.item]
+        var contentConfiguration = cell.defaultContentConfiguration()
+        contentConfiguration.text = reminder.title
+        contentConfiguration.secondaryText = reminder.dueDate.dayAndTimeText
+        contentConfiguration.secondaryTextProperties.font = UIFont.preferredFont(forTextStyle: .caption1)
+//        contentConfiguration.secondaryText = reminder.notes
+        cell.contentConfiguration = contentConfiguration
+    }
+}
